Refactor checksummers to use `long` for polynomial representation.
Updated the polynomial type from `int` to `long` across checksummer implementations to ensure compatibility with larger polynomial values. Adjusted comments and calculations to align with the new datatype changes.

+++ src/main/java/io/github/mtrevisan/boxon/annotations/checksummers/BSD16.java
@@ -41,8 +41,8 @@ public final class BSD16 implements Checksummer{
 	}
 
 	@Override
-	public int getPolynomial(){
-		return 0;
+	public long getPolynomial(){
+		return 0l;
 	}
 
 	@Override
